notification_service/notification/requirements.txt
Django==4
psycopg2-binary==2.9.3
python-dotenv==0.20.0
gunicorn==20.1.0
django-split-settings==1.1.0
celery==5.2.7
django-celery-beat==2.3.0
django-celery-results==2.2.0
celery_once==3.0.1
sendgrid==6.9.7
djangorestframework==3.13.1
django-filter==22.1
drf-yasg==1.21.3
pydantic==1.10.0